refactor(opencode): retire experimental workspace Hono routes
