--- packages/opencode/src/server/routes/control/workspace.ts
@@ -1,1 +1,0 @@
-export { WorkspaceRoutes } from "../../instance/workspace"
