--- packages/opencode/src/server/routes/instance/experimental.ts
@@ -1,1 +1,0 @@
-export { ExperimentalRoutes } from "../../instance/experimental"
